feat: Implement various hooks for user functionalities including home dashboard, login, logout, office attendance, reminders, and room bookings
- Added `useHomeDashboard` for managing calendar events and room bookings.
- Introduced `useLoginForm` for handling login form state and validation.
- Created `useLogoutWithConfirmation` for logout functionality with user confirmation.
- Developed `useOfficeAttendance` to track user attendance status.
- Implemented `useReminders` for fetching and managing user reminders.
- Added `useUserRoomBookings` to retrieve room booking summaries for users.
- Created `useRoomsAdmin` for managing room data including creation, editing, and deletion.
- Introduced `useSidebar` for sidebar state management and logout handling.
- Defined types for attendance, events, reminders, rooms, and users to ensure type safety.
- Added utility functions for date formatting to standardize date display across the application.
