+++ src/client/src/constants/calendar_constants.ts
@@ -1,0 +1,11 @@
+export const MONTH_NAMES = [
+  'January', 'February', 'March', 'April', 'May', 'June',
+  'July', 'August', 'September', 'October', 'November', 'December'
+];
+
+export const WEEKDAY_LABELS = ['Sun', 'Mon', 'Tue', 'Wed', 'Thu', 'Fri', 'Sat'];
+
+// Conservative estimate for event card height: date badge + title + time + description (2 lines) + attending + padding
+export const ESTIMATED_EVENT_CARD_HEIGHT = 115;
+export const MIN_UPCOMING_EVENTS = 1;
+export const DEFAULT_UPCOMING_EVENTS = 3;
