+++ src/client/src/hooks/login_form_hooks.tsx
@@ -1,0 +1,170 @@
+import { useState, useCallback } from 'react';
+import { useNavigate } from 'react-router-dom';
+import { useAuth } from '../auth/AuthContext';
+import { apiFetch } from '../config/api';
+
+/*
+ ====================================
+LOGIN FORM HOOKS
+ ====================================
+ */
+
+/*
+ Custom hook for form validation and error handling
+ */
+export const useFormValidation = () => {
+  const [error, setError] = useState<string | null>(null);
+
+  const clearError = useCallback(() => setError(null), []);
+
+  const validateEmail = useCallback((email: string): boolean => {
+    if (!email) {
+      setError('Email is required.');
+      return false;
+    }
+    if (!/^[^@\s]+@[^@\s]+\.[^@\s]+$/.test(email)) {
+      setError('Enter a valid email address.');
+      return false;
+    }
+    return true;
+  }, []);
+
+  const validatePassword = useCallback((password: string, minLength: number = 8): boolean => {
+    if (!password) {
+      setError('Password is required.');
+      return false;
+    }
+    if (password.length < minLength) {
+      setError(`Password must be at least ${minLength} characters long.`);
+      return false;
+    }
+    return true;
+  }, []);
+
+  const validatePasswordMatch = useCallback((password: string, confirmPassword: string): boolean => {
+    if (password !== confirmPassword) {
+      setError('Passwords do not match.');
+      return false;
+    }
+    return true;
+  }, []);
+
+  const validateRequired = useCallback((fields: Record<string, any>, fieldNames?: string[]): boolean => {
+    const fieldsToCheck = fieldNames || Object.keys(fields);
+    const emptyFields = fieldsToCheck.filter(key => !fields[key]);
+    
+    if (emptyFields.length > 0) {
+      setError('All fields are required.');
+      return false;
+    }
+    return true;
+  }, []);
+
+  return {
+    error,
+    setError,
+    clearError,
+    validateEmail,
+    validatePassword,
+    validatePasswordMatch,
+    validateRequired,
+  };
+};
+
+/*
+ Custom hook for password visibility toggle
+ */
+export const usePasswordVisibility = () => {
+  const [showPassword, setShowPassword] = useState(false);
+
+  const togglePasswordVisibility = useCallback(() => {
+    setShowPassword(prev => !prev);
+  }, []);
+
+  return {
+    showPassword,
+    setShowPassword,
+    togglePasswordVisibility,
+  };
+};
+
+/*
+ Custom hook for login form logic
+ */
+export const useLoginForm = () => {
+  const [email, setEmail] = useState('');
+  const [password, setPassword] = useState('');
+  const [loading, setLoading] = useState(false);
+  const { error, setError, validateEmail, validatePassword, clearError } = useFormValidation();
+  const { showPassword, togglePasswordVisibility } = usePasswordVisibility();
+  const navigate = useNavigate();
+  const { login } = useAuth();
+
+  const validate = useCallback((): boolean => {
+    clearError();
+    if (!email || !password) {
+      setError('Email and password are required.');
+      return false;
+    }
+    if (!validateEmail(email)) {
+      return false;
+    }
+    return true;
+  }, [email, password, validateEmail, clearError, setError]);
+
+  const handleSubmit = useCallback(async (e: React.FormEvent) => {
+    e.preventDefault();
+    if (!validate()) return;
+    setLoading(true);
+    setError(null);
+
+    try {
+      const response = await apiFetch('/api/auth/login', {
+        method: 'POST',
+        body: JSON.stringify({ email, password })
+      });
+
+      if (!response.ok) {
+        const text = await response.text();
+        throw new Error(text || 'Login failed');
+      }
+
+      const data = await response.json();
+      // Expected format : { token: string, user: { userId, name, email, role } }
+      const token: string | undefined = data.token || data.Token; 
+      const user = data.user;
+
+      if (!token || !user) {
+        throw new Error('Malformed login response');
+      }
+
+
+      // Persist token & user via context
+      login({
+        userId: user.userId,
+        name: user.name,
+        email: user.email,
+        role: user.role,
+      }, token);
+
+      navigate('/home');
+    } catch (err: any) {
+      console.error('Login error:', err);
+      setError(err.message ?? 'Unexpected error during login');
+    } finally {
+      setLoading(false);
+    }
+  }, [email, password, validate, login, navigate, setError]);
+
+  return {
+    email,
+    setEmail,
+    password,
+    setPassword,
+    loading,
+    error,
+    showPassword,
+    togglePasswordVisibility,
+    handleSubmit,
+  };
+};
