+++ src/client/src/hooks/sidebar_hooks.tsx
@@ -1,0 +1,46 @@
+import { useState, useCallback, useEffect } from 'react';
+import { useNavigate } from 'react-router-dom';
+import { useAuth } from '../auth/AuthContext';
+
+/*
+ ====================================
+SIDEBAR HOOKS
+ ====================================
+ */
+/*
+ Custom hook for sidebar logic
+ */
+export const useSidebar = () => {
+  const [isCollapsed, setIsCollapsed] = useState(() => {
+    const saved = localStorage.getItem('sidebarCollapsed');
+    return saved === 'true';
+  });
+  const navigate = useNavigate();
+  const { logout } = useAuth();
+
+  const toggleSidebar = useCallback(() => {
+    setIsCollapsed(prev => {
+      const newValue = !prev;
+      localStorage.setItem('sidebarCollapsed', String(newValue));
+      return newValue;
+    });
+  }, []);
+
+  const handleLogout = useCallback(() => {
+    const confirmed = window.confirm('Are you sure you want to logout?');
+    if (confirmed) {
+      logout();
+      navigate('/login');
+    }
+  }, [logout, navigate]);
+
+  useEffect(() => {
+    document.body.classList.toggle('sidebar-collapsed', isCollapsed);
+  }, [isCollapsed]);
+
+  return {
+    isCollapsed,
+    toggleSidebar,
+    handleLogout,
+  };
+};
