+++ src/client/src/types/attendance_types.ts
@@ -1,0 +1,6 @@
+/**
+ * Office attendance types
+ * Used for tracking employee office presence status
+ */
+
+export type AttendanceStatus = 'Present' | 'Absent' | 'Remote';
